efiboot/x86: Experimental support to memory mapped io for raw com
Tested by msaitoh@ for an Elkhart Lake machine.

XXX
Pass parameters to kernel.

+++ sys/arch/i386/stand/efiboot/efiboot.h
@@ -70,7 +70,7 @@ int utf8_to_ucs2(const char *, CHAR16 **, size_t *);
 
 /* eficons.c */
 int cninit(void);
-void efi_consinit(int, int, int);
+void efi_consinit(int, u_long, int);
 void efi_cons_show(void);
 void command_text(char *);
 void command_gop(char *);
