+++ sys/arch/i386/stand/efiboot/eficpufunc.h
@@ -34,5 +34,5 @@
  * Currently, only used for raw console I/O by eficons.c.
  */
 
-uint8_t	inb(uint32_t);
-void	outb(uint32_t, uint8_t);
+uint8_t	inb(u_long);
+void	outb(u_long, uint8_t);
